app/requirements.txt
Flask
Flask-HTTPAuth
SQLAlchemy
psycopg2-binary
PyJWT
argon2-cffi
PyYAML
